Bug 1389844 - P3. Handle case where currentTime isn't contained in buffered range. r=gerald
When playback starts, currentTime is always 0, and even if the buffered data doesn't contain currentTime it is possible for playback to progress as we always allow up to 500ms gap in the buffered data.

As such, we must use fuzzing on the interval's start time when determining if we have future data.

MozReview-Commit-ID: Ki9QxmKhfdY

UltraBlame original commit: 45a51fab5dbf2d91e902c905e803ae73f2930d59