Bug 1388660 - part1 : AudioData::IsAudible() should return false when its data is super small and near to zero. r=jwwang
When the value of data is too small to be heard, AudioData::IsAudible() should return false so that we won't show the sound indicator for silent media.

In this case, the loudness of reported video is -673 dBFS, it's impossible to be heard.

MozReview-Commit-ID: Ewiko7RpkeX

UltraBlame original commit: 91caacf94f7528d724054e2f18516da13d2a5a71