Bug 1393128 - 3. Initialize DownloadNotifications outside of browser.js; r=esawin
In order to show download notifications, we need to initialize the
DownloadNotifications module outside of browser.js. This patch moves
initialization to BrowserCLH.js, and includes a refactoring of the
`addObserverScripts` function. The "chrome-document-interactive"
notification is used to trigger initialization because it is roughly
equivalent to where we used to initialize the module inside browser.js.

MozReview-Commit-ID: 8o1KZWRt69K

UltraBlame original commit: b89c2fffc332bfe98dcb669bd8a8168c3b3c8275