Bug 1380252 - Send only the crashing thread's stack in crash pings; r=ted.mielczarek
This modifies the minidump analyzer to only output the crashing thread's stack
when invoked. This reduces the raw size of the crash ping by 50% to 70%
depending on the program state when the crash occurred. The minidump-analyzer
can still produce the complete output when invoked with the '--full' option.

MozReview-Commit-ID: 8KaITWqfBKR

UltraBlame original commit: 022687c2669f744d4cfed3765f18038da4edcce5