Bug 1395469 - Check discarded ChromeWindow in staleness test. r=me
HTML elements' ownerGlobal is a WindowProxy that indirects operations
to the current browsing context.  However, for XUL elements this global
is ChromeWindow which gets reset to null when the window is discarded.

For the same reason, since win indirects to the current browsing context's
associated window, we need to test that the element's node document is
not the active document.

Finally test that its shadow-including root is a document.

MozReview-Commit-ID: EYh18P8DbcN

UltraBlame original commit: f2124b88c59d4b65b1a8dea5e33813d6ff5dbf9d