Bug 1330412 - Stylo seq. runs for memory / perf only. r=jmaher
We no longer need to run unit / integration tests on Stylo sequential, but we'd
like to retain it for memory and performance checks.

In addition, we no longer need to explicitly `--enable-stylo` since it is the
default.  This also drops the `stylo` tag in Perfherder.

MozReview-Commit-ID: 3OI8Y7Q5PW

UltraBlame original commit: d37df1ba1ca94a3a073266e25bd0e9e6001be9d5