Bug 1395752 - Define and use sparse profile for Sphinx docs; r=dustin
The Sphinx documentation only needs access to a relatively small number
of files in the repo in order to be generated. It is a good candidate
for using sparse profiles.

This commit defines and uses a "sphinx-docs" sparse profile containing
only the files relevant to Sphinx documentation generation.

There are some quirks with the profile:

* All moz.build files are included. This bloats the profile
  by >1000 files. Worse, it realizes directories that have no business
  being realized. This clutters the checkout and makes it harder to
  find things. There is a moz.build reader that knows how to retrive
  file data from Mercurial. We could use that. This feels like follow-up
  fodder.

* All mach_commands.py files are included. `mach help` says you can do
  things that you aren't able to do in the sparse checkout. There isn't
  a good way to add all *.py files while excluding mach_commands.py
  files. We /could/ do it with regular expressions. But those are slow.
  Let's leave it as is for now and come up with a better solution later.

MozReview-Commit-ID: 7yiqGGE1nAh

UltraBlame original commit: 5ecc8d4b725a2974260eb5bef495b91771fadbb0