Bug 1394381 - Add logging output for Marionette connection attempts. r=maja_zf
If Marionette client cannot connect to the server, the attempts
have to be logged to stderr. This allows us to inspect the reason
why a connection from the client could not be made, and that it
still tries to connect (see bug 1362293).

MozReview-Commit-ID: ElE1M73ums8

UltraBlame original commit: 1e12a866239c8ab6d5ee68c31731d66a54733d81