Bug 1387130 - Use original tabstrip scrolling behaviour when using scrollbuttons. r=dao
In bug 1356705, we switched scrollbox to use CSS smooth scroll when
the scrollbox is configured to scroll smoothly. This caused the tab
strip to scroll with a "pulse" when using the arrow scrollbuttons.
This is because we scroll by a single tab each time, as opposed to
scrolling by pixels.

This reverts part of bug 1356705 so that we use instant scrolling
instead of smooth scrolling in the scrollbuttons case, which returns
the original behaviour of the strip without the pulse.

MozReview-Commit-ID: D8QQ8kQ7AjM

UltraBlame original commit: ad5c7dbaa4d6c2f31c3b5a89c159f781a2f92902