Bug 1398019 - Favicons missing in Bookmarks toolbar on startup. r=past
This is due to the fact now we set the icons only when the nodes visibility is updated.
The patch always updates nodes visibility when we rebuild the toolbar.
updateChevron() is being renamed accordingly and acts regardless of chevron.collapsed.

MozReview-Commit-ID: Cz1U710J42M

UltraBlame original commit: 0a49941e45a5d58b01051b4283d4bd1fff7aadd1