Bug 1396076 - Part 3 - Switch to a LayoutChangeListener for setting up the delegates. r=jwu
When the device is rotated between portrait and landscape mode (or vice versa), the View's final size is not yet available during the first OnPreDrawListener call for some of the TabsLayoutItemViews. It would be possible to skip the listener removal, so this is no longer used as a one-shot listener, however onPreDraw is frequently called even when the View's dimensions haven't changed.
Therefore, we switch this to an OnLayoutChangeListener, which is more appropriate for what we need.

MozReview-Commit-ID: 6JAzXdEBxbL

UltraBlame original commit: 60f7b76c8f7f7d3c00a2f7be03aebe23b90b3ae2