Bug 1388396: Increase max favicon scale factor to 3. r=sebastian
This is part 1 of the good-enough approach: see the MAX_SCALE_FACTOR comment
as to what we're aiming for. The next changeset will discard any icons that
do not look good being scaled so much.

The MAX_SCALE_FACTOR field is non-private because it's used in the next
changeset.

MozReview-Commit-ID: HGzdQBEuMAy

UltraBlame original commit: 0aff84a0a8af1d23f1cd8f9abd810f82874731cd