Bug 1216885 - Make nsISelectionPrivate not inherit from nsISelection. r=smaug
nsISelectionPrivate is accessible to script, while nsISelection is
not, so making the former inherit from the latter means script doesn't
have a complete view of the inheritance chain so the XPIDL compiler
produces an error.

It turns out that nothing in script relies on this inheritance, which
makes sense because I'm not sure how it would even work, so just
remove it.

MozReview-Commit-ID: 3Py2T7cprlD

UltraBlame original commit: 1ca25d95e1aac287652580b29e3220201c1d6086