Bug 1393687 - Use VideoAdapter to handle max-fs; r=jesup
This uses the VideoAdapter to handle the max-fs constraint rather than
attempting to calculate and enforce it manually within the VideoConduit.
The resolution is taken as the small of the max-fs constraint and the
latest request from the video sink.

The old unit tests for max-fs and max-fr are removed as the size
calculation has now been delegated to the VideoAdapter and the tests do
not verify whether or not video has actually been scaled.

MozReview-Commit-ID: 18khmiNageR

UltraBlame original commit: 17637a9cb3db623a41e2dc29a0f56f9a2cefa35f