Bug 1387594 - Sprinkle -moz-font-smoothing-background-color declarations over the CSS. r=dao
There's only one interesting case here: the active tab. When rendering the label
of an overflowing active tab into the fadeout mask surface, text rendering must
not use the font smoothing background color for dark vibrancy. Instead, it needs
to use the color of the tab itself for best results.
Alternatively, we could set the font smoothing background color of the active
tab to "transparent", because this text is not on top of a vibrant background.
However, doing so would lose the subpixel AA and would not look as crisp.

MozReview-Commit-ID: 28MKCz1vmb9

UltraBlame original commit: 3dcafbdc27cf956aaddf2a0997c8c2ac02829b28