Bug 1399460 - Fix for browser_webconsole_check_stubs_console_api. r=Honza
The test that checked console api stubs was failing because we were trying
to reach the server while the connection was closed, which throws an error.
This patch filters-out log messages in the console output so we don't get
unwanted server communication. The output is not important here since we
are only listening to the consoleAPICall event.

MozReview-Commit-ID: Ay5fQ1CSbLx

UltraBlame original commit: 8b0058e2a4a8a123e09470b7ee975963be88a158