Bug 1366989 - Part 1: Avoid to refresh whole panel. r=pbro
Currently the animation inspector re-generates the entire animation timeline
whenever an animation is added, changed, etc.
To avoid this, averts to re-render the component which no needs.

In this implementation, premises the actorID can be used as unique id for each
animations. The mechanism is below.

At initial time, renders all actors  as normally. In this time, holds actorID
and related components to componentsMap.
Next, in case of that needs to update the UI, gets animation actors from server,
and compares actorID of both the actors and componentsMap. If retrieved actorID
exists in componentsMap, updates the view area only without re-rendering.
For example, supposes, has an animation (actid-1) when opens the inspector, and
a new animation (actid-2) was added a little later.
At initial rendering, holds "actid-1” of first animation as key and related
components to componentsMap. Next, when “actid-2” animation is added to document,
can get animation actors that are “actid-1” and “actid-2” from server. Because
“actid-1” is already held in componentsMap, updates “actid-1”’s view area. This
is because TimeScale will be updated. Then "actid-2” render as normally since
componentMap does not have the actorID. After rendered, holds “actid-2” and
related components.

However, even if actorID exists, if keyframes (tracks) and effect timing
(state) differ, re-render that. Also, if iterationCount of effect timing
represents Infinity, do re-rendering. Because the display area expands by the
end of the currently displayed time.

And, if actorID in componentsMap is not in retrieved actors, removes related
components.

MozReview-Commit-ID: GmifRX3GzYd

UltraBlame original commit: f391fb90181c5d230f3a8052829727e019be082a