Bug 1371849 - remove style editor's source-map pref; r=gl
This changes the style editor and inspector to use the same pref that
the other tools are using.  This pref is a bit of a misnomer for now,
because the inspector and style editor aren't actually using the
client-side service; but I think the name is not very important, and
unifying the prefs is worthwhile.

MozReview-Commit-ID: FlwCICnoaX4

UltraBlame original commit: 9393990862b0702cdc4c6ae0afa68daa4c1faedf